Obtain time series of EV connected from sessions data set

Obtain time series of EV connected from sessions data set

Usage

get_n_connections(
  sessions,
  dttm_seq = NULL,
  by = "Profile",
  resolution = 15,
  align_time = FALSE,
  mc.cores = 1
)

Arguments

sessions

tibble, sessions data set in standard format marked by {evprof} package (see this article)

dttm_seq

sequence of datetime values that will be the datetime variable of the returned time-series data frame.

by

character, being 'Profile' or 'Session'. When by='Profile' each column corresponds to an EV user profile.

resolution

integer, time resolution (in minutes) of the sessions datetime variables. If dttm_seq is defined this parameter is ignored.

align_time

logical, whether to align time variables or sessions with the corresponding time resolution

mc.cores

integer, number of cores to use. Must be at least one, and parallelization requires at least two cores.

Value

tibble

Details

Note that the time resolution of variable ConnectionStartDateTime must coincide with resolution parameter. For example, if a charging session in sessions starts charging at 15:32 and resolution = 15, the load of this session won't be computed. To solve this, the parameter align_time = TRUE would align charging sessions' start time according to resolution, so following the previous example the session would start at 15:30.

Examples

library(lubridate)
library(dplyr)

# Get occupancy with the complete datetime sequence from the sessions
sessions <- head(evsim::california_ev_sessions, 100)
connections <- get_n_connections(
  sessions,
  by = "Session",
  resolution = 60,
  align_time = TRUE
)
connections %>% plot_ts()

# Get occupancy with a custom datetime sequence and resolution of 15 minutes
sessions <- head(evsim::california_ev_sessions_profiles, 100)
dttm_seq <- seq.POSIXt(
  as_datetime(dmy(01102018)) %>% force_tz(tz(sessions$ConnectionStartDateTime)),
  as_datetime(dmy(07102018)) %>% force_tz(tz(sessions$ConnectionStartDateTime)),
  by = "15 mins"
)
connections <- get_n_connections(
  sessions,
  dttm_seq = dttm_seq,
  by = "Profile",
  resolution = 15,
  align_time = TRUE
)
connections %>% plot_ts()